+++ server/src/nest/admin/admin.bridge.ts
@@ -5,7 +5,6 @@ import { AddonsService } from '../addons/addons.service';
 import { SettingsService } from '../settings/settings.service';
 import { AuthService } from '../auth/auth.service';
 import { PasskeyService } from '../auth/passkey.service';
-import { AtlasService } from '../atlas/atlas.service';
 import { UserCleanupService } from '../auth/user-cleanup.service';
 import { WebauthnConfigService } from '../auth/webauthn-config.service';
 import { TripMembershipService } from '../trip-membership/trip-membership.service';
@@ -41,7 +40,7 @@ const webauthn = new WebauthnConfigService(dbs);
 const userCleanup = new UserCleanupService(dbs);
 const mailer = new MailerService(dbs);
 const notifPrefs = new NotificationPreferencesService(dbs, mailer);
-const auth = new AuthService(dbs, permissions, new AtlasService(dbs), new TripMembershipService(dbs), webauthn, userCleanup, mailer);
+const auth = new AuthService(dbs, permissions, new TripMembershipService(dbs), webauthn, userCleanup, mailer);
 const admin = new AdminService(
   dbs,
   new AddonsService(dbs),
