+++ server/src/nest/atlas/atlas.module.ts
@@ -1,5 +1,6 @@
 import { Module } from '@nestjs/common';
 import { AtlasController } from './atlas.controller';
+import { TravelStatsController } from './travel-stats.controller';
 import { AtlasService } from './atlas.service';
 import { AtlasRpc } from './atlas.rpc';
 import { PluginGuardsModule } from '../plugins/host/plugin-guards.module';
@@ -10,10 +11,14 @@ import { AtlasMcp } from './atlas.mcp';
  * AtlasService for the plugin RPC surface (AtlasRpc injects it).
  * AtlasMcp is a provider (not a controller) — the nest-mcp registry discovers
  * it after app.init().
+ *
+ * TravelStatsController serves GET /api/auth/travel-stats. The path stays where
+ * the client expects it; the code sits here because getTravelStats reads Atlas
+ * data. See the comment in that file.
  */
 @Module({
   imports: [PluginGuardsModule],
-  controllers: [AtlasController],
+  controllers: [AtlasController, TravelStatsController],
   providers: [AtlasService, AtlasMcp, AtlasRpc],
   exports: [AtlasService],
 })
