+++ server/src/nest/auth/auth.bridge.ts
@@ -1,7 +1,6 @@
 import { db } from '../../db/database';
 import { DatabaseService } from '../database/database.service';
 import { PermissionsService } from '../permissions/permissions.service';
-import { AtlasService } from '../atlas/atlas.service';
 import { TripMembershipService } from '../trip-membership/trip-membership.service';
 import { AuthService } from './auth.service';
 import { MailerService } from '../notifications/mailer/mailer.service';
@@ -15,10 +14,12 @@ import { User } from '../../types';
  * the legacy tool registrars (mcp/tools/journey|notifications|transports), the
  * not-yet-migrated services/adminService, and — as the one in-container
  * exception —
- * atlas.mcp.ts, which cannot inject AuthService without closing an
- * AuthModule↔AtlasModule cycle (AuthService injects AtlasService for
- * getTravelStats; same documented trade-off as places.mcp.ts keeping its
- * assignments.bridge imports).
+ * atlas.mcp.ts. The cycle that used to make that mandatory is gone
+ * (getTravelStats moved to AtlasService, so AuthModule no longer imports
+ * AtlasModule); what keeps it here is cost, not correctness: a direct import
+ * would drag the whole auth chain into every Atlas consumer for one
+ * isDemoUser call. Same open trade-off as places.mcp.ts keeping its
+ * assignments.bridge imports.
  *
  * Exports the legacy services/authService names 1:1 so repointing a consumer
  * is an import-path-only diff. Inside the container, inject AuthService
@@ -39,7 +40,6 @@ const permissions = new PermissionsService(dbs);
 const auth = new AuthService(
   dbs,
   permissions,
-  new AtlasService(dbs),
   new TripMembershipService(dbs),
   new WebauthnConfigService(dbs),
   new UserCleanupService(dbs),
